# apps/api-gateway/docs/openapi.json
{
  "components": {
    "responses": {
      "ForbiddenError": {
        "content": {
          "application/json": {
            "example": {
              "detail": "admin role required"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "User terautentikasi tetapi role-nya tidak memiliki akses."
      },
      "UnauthorizedError": {
        "content": {
          "application/json": {
            "example": {
              "detail": "Tidak terautentikasi"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Firebase ID token tidak ada, tidak valid, atau kedaluwarsa."
      }
    },
    "schemas": {
...
      "BatchRunResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "enqueued": {
            "title": "Enqueued",
            "type": "boolean"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "enqueued"
        ],
        "title": "BatchRunResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelRegistryRow": {
        "properties": {
          "artifact_path": {
            "title": "Artifact Path",
            "type": "string"
          },
          "feature_set_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Set Version"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "model_type": {
            "title": "Model Type",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "model_name",
          "model_type",
          "target_type",
          "version",
          "artifact_path",
          "is_active"
        ],
        "title": "ModelRegistryRow",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "FirebaseBearer": {
        "bearerFormat": "JWT",
        "description": "Firebase ID token. Masukkan token saja; Swagger UI otomatis menambahkan prefix `Bearer`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "team@inflasi.id",
      "name": "INFLASI Team",
      "url": "https://inflasi.id/"
    },
...
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "summary": "Food-inflation monitoring + forecasting backend (FastAPI + TimescaleDB + ml-gateway).",
    "title": "INFLASI API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/models": {
      "get": {
        "operationId": "list_models_api_admin_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "model_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Type"
            }
          },
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Type"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizon"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelRegistryRow"
                  },
                  "title": "Response List Models Api Admin Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "List Models",
        "tags": [
          "admin-models"
        ]
      },
      "post": {
        "operationId": "register_model_api_admin_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRegistryRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Register Model",
        "tags": [
          "admin-models"
        ]
      }
    },
    "/api/admin/models/active": {
      "get": {
        "operationId": "list_active_api_admin_models_active_get",
        "parameters": [
          {
            "in": "query",
            "name": "model_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Type"
            }
          },
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Type"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizon"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelRegistryRow"
                  },
                  "title": "Response List Active Api Admin Models Active Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "List Active",
        "tags": [
          "admin-models"
        ]
      }
    },
    "/api/admin/models/batch-run": {
      "post": {
        "description": "Enqueue a `batch_forecast` run on the Redis stream.\n\nReturns 202-ish semantics: the request just publishes — the worker\nprocesses asynchronously. Inspect `job:batch_forecast:{job_id}` for status.",
        "operationId": "trigger_batch_forecast_api_admin_models_batch_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Trigger Batch Forecast",
        "tags": [
          "admin-models"
        ]
      }
    },
    "/api/admin/models/batch-run/{job_id}": {
      "get": {
        "operationId": "get_batch_run_status_api_admin_models_batch_run__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Batch Run Status Api Admin Models Batch Run  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Get Batch Run Status",
        "tags": [
          "admin-models"
        ]
      }
    },
    "/api/admin/models/retrain": {
      "post": {
        "description": "Enqueue a retraining run for `model_type` (and optional horizon).\n\nSpawns the trainer subprocess inside the api-gateway pod. Audit row is\nwritten to `model_training_runs`; status mirror lives in\n`job:retrain:{run_id}` once the worker picks it up.",
        "operationId": "trigger_retrain_api_admin_models_retrain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrainEnqueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Trigger Retrain",
        "tags": [
          "admin-models"
        ]
      }
    },
    "/api/admin/models/training-runs": {
      "get": {
        "operationId": "list_training_runs_api_admin_models_training_runs_get",
        "parameters": [
          {
            "description": "RUNNING | SUCCESS | FAILED | SKIPPED",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "RUNNING | SUCCESS | FAILED | SKIPPED",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrainingRunRow"
                  },
                  "title": "Response List Training Runs Api Admin Models Training Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "List Training Runs",
        "tags": [
          "admin-models"
        ]
      }
    },
    "/api/admin/models/training-runs/{run_id}/job-status": {
      "get": {
        "operationId": "get_training_run_job_status_api_admin_models_training_runs__run_id__job_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Training Run Job Status Api Admin Models Training Runs  Run Id  Job Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Get Training Run Job Status",
        "tags": [
          "admin-models"
        ]
      }
    },
    "/api/admin/models/{model_id}/promote": {
      "post": {
        "operationId": "promote_model_api_admin_models__model_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRegistryRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Promote Model",
        "tags": [
          "admin-models"
        ]
      }
    },
...
    "/api/internal/models": {
      "post": {
        "description": "Register an artifact produced by a Kubernetes training Job.",
        "operationId": "register_model_from_training_api_internal_models_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Training-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Training-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRegistryRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Model From Training",
        "tags": [
          "internal"
        ]
      }
    },
    "/api/internal/models/active": {
      "get": {
        "description": "Return the single active model row for the given slot.\n\nReturns 404 when no active model is registered; callers should treat that as\n\"this model isn't available yet\" and degrade gracefully.",
        "operationId": "get_active_model_api_internal_models_active_get",
        "parameters": [
          {
            "description": "lightgbm | prophet | sarimax | tft | ensemble | stacking",
            "in": "query",
            "name": "model_type",
            "required": true,
            "schema": {
              "description": "lightgbm | prophet | sarimax | tft | ensemble | stacking",
              "title": "Model Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "default": "price",
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "description": "Required for horizon-scoped models (lightgbm/tft)",
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required for horizon-scoped models (lightgbm/tft)",
              "title": "Horizon"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Active Model Api Internal Models Active Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Active Model",
        "tags": [
          "internal"
        ]
      }
    },
...
    "/api/users/admin/{user_id}": {
      "patch": {
        "operationId": "admin_update_user_api_users_admin__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Admin Update User",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Server saat ini (relative)",
      "url": "/"
    },
    {
      "description": "Local dev",
      "url": "http://localhost:8000"
    },
    {
      "description": "Production",
      "url": "https://api.inflasi.id"
    }
  ],
...
}